update row count
